split to multiple walker
